refactor(onboarding): make IntegrationsDrawer reusable and use it in stories
The integrations drawer isn't onboarding-specific, so move it out of the
GettingStarted folder to a shared components/IntegrationsDrawer/ module
(carrying its integrationsCatalog + generated guides, re-exported via index.ts
so other surfaces can reuse it).

Wire the OnboardingAccordion "Browse integration guides" / "Language SDKs"
links to the real IntegrationsDrawer instead of the previous mock docs drawer,
opening the languages category from the SDKs link. Removes the now-unused mock
(IntegrationsDocsDrawer/DocRow and its static lists).
