+++ packages/app/src/components/IntegrationsDrawer/index.ts
@@ -1,3 +1,4 @@
+export * from './integrationsCatalog';
 export {
   IntegrationsDrawer,
   type IntegrationsDrawerProps,
